# Improving WASM instance spawn performance