--- a/core/wasm/src/cache.rs
+++ b/core/wasm/src/cache.rs
@@ -1,13 +1,26 @@
-use tempfile::TempDir;
-use std::sync::Mutex;
-use std::path::{Path, PathBuf};
+use cached::SizedCache;
+use wasmer_runtime;
 
-lazy_static! {
-    static ref CACHE_TEMP_DIR: Mutex<TempDir> =
-        Mutex::new(TempDir::new()
-            .expect("create wasm cache temp directory"));
-}
+use primitives::traits::Encode;
+use primitives::hash::hash;
+use crate::prepare;
+use crate::types::{Config, Error};
 
-pub(crate) fn get_cached_path<P: AsRef<Path>>(filename: P) -> PathBuf {
-    CACHE_TEMP_DIR.lock().unwrap().path().join(filename)
+/// Cache size in number of cached modules to hold.
+const CACHE_SIZE: usize = 1024;
+
+cached_key_result! {
+    MODULES: SizedCache<String, wasmer_runtime::Cache> = SizedCache::with_size(CACHE_SIZE);
+    Key = {
+        format!("{}:{}",
+            hash(code),
+            hash(&config.encode().expect("encoding of config shouldn't fail")),
+        )
+    };
+
+    fn compile_cached_module(code: &[u8], config: &Config) -> Result<wasmer_runtime::Cache, Error> = {
+        let prepared_code = prepare::prepare_contract(code, config).map_err(Error::Prepare)?;
+
+        wasmer_runtime::compile_cache(&prepared_code).map_err(|e| Error::Wasmer(e.into()))
+    }
 }