--- a/core/wasm/src/lib.rs
+++ b/core/wasm/src/lib.rs
@@ -1,6 +1,9 @@
 extern crate parity_wasm;
 extern crate pwasm_utils;
 extern crate wasmi;
+extern crate tempfile;
+#[macro_use]
+extern crate lazy_static;
 
 #[cfg(test)]
 #[macro_use]
@@ -14,6 +17,7 @@ extern crate primitives;
 #[macro_use]
 extern crate log;
 
+mod cache;
 pub mod executor;
 pub mod ext;
 mod prepare;