- Standardize memory so we can initiate memory without code inspection
- Adding file cache that stores compiled code into temporary directory

Caveats:
- Disk Cache is insecure.
- It uses lazy static init, that might crash if temp dir is failed to create.
- Wasmer cache has one of the functions as unsafe. We should probably change this cache to be LRU memory cache. I'll explore it next.
- It also has assumption about config for execution being the same.

But this change improves WASM benchmarks from 93ms overhead to much less:
```
test runtime_send_money                 ... bench:     541,875 ns/iter (+/- 225,776)
test runtime_wasm_benchmark_10_reads    ... bench:   1,941,265 ns/iter (+/- 298,672)
test runtime_wasm_benchmark_sum_1000    ... bench:   1,877,210 ns/iter (+/- 502,444)
test runtime_wasm_benchmark_sum_1000000 ... bench:  24,895,845 ns/iter (+/- 3,931,008)
test runtime_wasm_set_value             ... bench:   1,758,437 ns/iter (+/- 588,113)
```